changes to update user profile

+++ backend/routes/users/updateUser.js
@@ -1,13 +1,15 @@
 const Schemas = require("../../models/index");
 const logger = require("../../helpers/logger");
-
+const uploadImage = require("../../helpers/uploadImage");
 const deleteImage = require("../../helpers/deleteImage");
 module.exports = async (req, res) => {
   const ip = req.header("x-forwarded-for") || req.connection.remoteAddress;
-  const userId = req.params.id;
+  const userId = req.user._id;
   const name = req.body.name;
   const description = req.body.description;
   const imageData = req.files;
+  const deleteImageUrl = req.body.deleteImageUrl || false;
+
   //checking if the userId length is proper or not
   if (userId.length != 24) {
     return res.status(400).json({
@@ -16,7 +18,7 @@ module.exports = async (req, res) => {
     });
   }
   //checking if all the params exist
-  if (!name || !imageData || !description) {
+  if (!name && !imageData && !description) {
     return res.status(400).json({
       success: false,
       message: "Nothing to update!!",
@@ -47,12 +49,12 @@ module.exports = async (req, res) => {
 
     // if imageData is passed from the body then we will change the value of the user imageUrl in the database
 
-    if (imageData.length == "0") {
+    if (deleteImageUrl) {
       deleteImage(user.imageUrl);
       user.imageUrl = "";
-    } else if (imageData.length > 0) {
+    } else if (!deleteImageUrl && imageData.length > 0) {
       deleteImage(user.imageUrl);
-      const url = await uploadImage(imageData[0], `./user/${userId}`);
+      const url = await uploadImage(imageData[0], `/user/${userId}`);
       //Save url in database
       user.imageUrl = url;
     }
